requirements.txt
fastapi>=0.110,<0.128
uvicorn>=0.27,<0.40
python-dotenv>=1.0,<2.0
jinja2>=3.1,<4.0
beautifulsoup4>=4.12,<5.0
slowapi>=0.1,<0.2
python-multipart>=0.0.9,<0.1
starsessions>=1.3,<2.2
httpx>=0.27,<0.29
werkzeug>=3.0,<4.0
cuid2>=2.0,<3.0
nanoid>=2.0,<3.0
python-ulid>=2.7,<3.1
caspian-utils==0.0.32
tailwind-merge>=0.3.3,<0.4.0
